Decouple MainActivity & MainScreen
